Fix issues spotted by static analysis
- Fix condition in Colour_Generator::PickEndColours() to correctly check that both m_propcolours[0] and m_propcolours[1] are non-zero before assigning color flows
- Fix loop condition in Matter_Overlap::SelectB() by changing loop variable to signed int to prevent infinite loop
- Fix charge comparison in Single_Virtual_Correction::ComputeChargeFactors() to correctly compare particles `i` and `j`


(cherry picked from commit c19d9574b58aeebd314848cb85452dcd466224d0)

49d4c0b3 Fix charge comparison in Single_Virtual_Correction::ComputeChargeFactors() to...
e3151d1a Fix loop condition in Matter_Overlap::SelectB() by changing loop variable to...
ccf5fa56 Fix condition in Colour_Generator::PickEndColours() to correctly check that...

+++ AMISIC++/Tools/Matter_Overlap.C
@@ -45,7 +45,7 @@ double Matter_Overlap::SelectB(const bool & mode) const {
   // 1. select a radius R according to matter content:
   // 2. Select b according to d^2b O(b) = d b^2 exp(-b^2/R^2).
   double b, radius, rand = ran->Get();
-  for (size_t i=3;i>=0;i--) {
+  for (int i=3;i>=0;i--) {
     rand -= m_fraction[i];
     if (rand<=0.) { radius = m_radius[i]; break; }
   }
